pyproject.toml
[tool.poetry]
name = "articles-crew"
version = "0.1.0"
description = ""
authors = ["ZTZ <tzamirov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
crewai = {extras = ["tools"], version = "^0.30.11"}
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.4"
agentops = "^0.1.11"
gpt4all = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
